Extract boon-reset-state-for-switchw for debugging
We effectively cannot attach the debugger to `boon-reset-state-for-switchw` when
it is defined inside the call to `add-hook` because it is added to the hook as
an anonymous lambda instead of being called by name. When the definition is
extracted to the top level, we can debug it more effectively because it is
called by name.