Add bindings to `boon-emacs.el` (#133)
* Use "@" instead of "x" for `boon-switch-mark'

"x" is shadowed by `boon-command-map'. "@" was the shortcut for
`boon-switch-mark' in version 1 of boon-emacs.el, so it feels like a natural
choice to return to that.

* Use "^" for `delete-indentation'

The default keybinding for this command is M-^.

* Use "\\" for `indent-region'

The default keybinding for this is C-M-\.

* Use "v"/"V" for `scroll-up-command'/`scroll-down-command'

The default keybindings for these are C-v and M-v.