Don't reset the last buffer state if we can't find it (#136)
* Define boon-reset-state-for-switchw at top level

Defining this function at top level makes it easier to attach the debugger or
override the definiton.

* Don't try to reset state if last window is undefined

`old-selected-window` may return a frame instead of a window, or `nil` if there
is not a previous window selection (e.g. at Emacs startup). In either of these
cases, trying to find the old buffer and reset its state would produce an error,
so do nothing instead.